% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kenya.R
\docType{data}
\name{kenya}
\alias{kenya}
\title{Kenya voter registration experiment}
\format{
A tibble with 1,989,680 observations and 11 variables:
\describe{
\item{treatment}{factor variable that indicates either the control group or some combination of
treatment factors: canvassing, SMS reminder, local administrator, local admin and
SMS, local admin and canvassing}
\item{poverty}{a double variable for the poverty rate of the voting block for those assigned to it}
\item{distance}{a double variable for average distance from polling location for those assigned to it}
\item{pop_density}{a double variable for the population density of individual's block}
\item{block}{a character variable for the individual's voting block ID}
\item{poll_station}{a character variable for the individual's polling station ID}
\item{rv13}{a double variable for the number of registered voters at polling location during 2013}
\item{reg_byrv13}{a double variable for registered voters at polling location during intervention
period divided by registered voters at polling location during 2013}
}
}
\source{
\url{https://pdfs.semanticscholar.org/be58/c86f2366d5281d4800411d2dfcecd1a5c16a.pdf?_ga=2.260073289.590691361.1594131808-281689301.1594131808}
}
\usage{
kenya
}
\description{
This dataset is from a study entitled \href{https://dx.doi.org/10.2139/ssrn.3520421}{"Electoral
Administration in Fledgling Democracies:Experimental Evidence from Kenya"}. The study worked with
Kenya's electoral commission in 1,674 communities by assigning individuals to either a control/status
quo group or one of several interventions to encourage voter registration: SMS reminders, a local
administrator at a polling station, canvassing, a local admin and canvassing, or a local admin and SMS.
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab kenya \cr
   Number of rows \tab 1672 \cr
   Number of columns \tab 9 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   character \tab 2 \cr
   factor \tab 1 \cr
   numeric \tab 6 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   block \tab 0 \tab 1 \tab 7 \tab 12 \tab 0 \tab 279 \tab 0 \cr
   poll_station \tab 0 \tab 1 \tab 7 \tab 7 \tab 0 \tab 1672 \tab 0 \cr
}


\strong{Variable type: factor}\tabular{lrrlrl}{
   skim_variable \tab n_missing \tab complete_rate \tab ordered \tab n_unique \tab top_counts \cr
   treatment \tab 0 \tab 1 \tab FALSE \tab 6 \tab can: 279, loc: 279, loc: 279, loc: 279 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   poverty \tab 0 \tab 1 \tab 0.43 \tab 0.11 \tab 0.18 \tab 0.35 \tab 0.43 \tab 0.49 \tab 0.90 \tab ▃▇▅▁▁ \cr
   distance \tab 0 \tab 1 \tab 20.93 \tab 16.24 \tab 0.04 \tab 10.20 \tab 17.26 \tab 27.34 \tab 179.02 \tab ▇▁▁▁▁ \cr
   pop_density \tab 0 \tab 1 \tab 0.00 \tab 0.00 \tab 0.00 \tab 0.00 \tab 0.00 \tab 0.00 \tab 0.02 \tab ▇▁▁▁▁ \cr
   mean_age \tab 8 \tab 1 \tab 41.63 \tab 2.78 \tab 24.02 \tab 39.82 \tab 41.59 \tab 43.42 \tab 50.07 \tab ▁▁▂▇▂ \cr
   reg_byrv13 \tab 4 \tab 1 \tab 0.02 \tab 0.04 \tab 0.00 \tab 0.00 \tab 0.00 \tab 0.02 \tab 0.73 \tab ▇▁▁▁▁ \cr
   rv13 \tab 0 \tab 1 \tab 539.23 \tab 429.35 \tab 0.00 \tab 273.00 \tab 442.00 \tab 675.00 \tab 6764.00 \tab ▇▁▁▁▁ \cr
}
}
\author{
David Kane
}
\keyword{datasets}
