% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes.R
\docType{data}
\name{nhanes}
\alias{nhanes}
\title{National Health and Nutrition Examination Survey}
\format{
A tibble with 10,000 observations and 15 variables:
\describe{
\item{gender}{character variable with values "male" and "female"}
\item{survey_year}{integer variable with the year the survey was conducted in. Most tests were
conducted across two years, so the earlier year is used.}
\item{age}{character variable for age in years}
\item{race}{character variable with values "white", "black", "hispanic", "mexican", and "other"}
\item{education}{character variable with values "middle school", "high school", "some college", "college"}
\item{income}{integer variable for household income groups ranked from 1 to 5}
\item{weight}{double variable for weight in kilograms}
\item{height}{double variable for height in centimeters}
\item{bmi}{double variable for body mass index}
\item{pulse}{integer variable for pulse in beats per minute or bpm}
\item{diabetes}{integer variable with 1 and 0 corresponding to having and not having diabetes, respectively}
\item{general_health}{integer variable for general health ranked from 1-5. 1 maps to "poor", 2 is "fair", 3 is "good", 4 is "very good", and 5 is "excellent".}
\item{depressed}{character variable answering question how often patient feels depressed.
Includes values "several", "none", "most"}
\item{pregnancies}{integer variable for number of pregnanices.
NA if the individual, regardless of gender, hasn't experienced a pregnancy}
\item{sleep_night_hrs}{integer value for hours of sleep per night on average}

}
}
\source{
\url{https://cran.r-project.org/web/packages/NHANES/NHANES.pdf}
}
\usage{
nhanes
}
\description{
See \href{https://www.cdc.gov/nchs/nhanes/index.htm}{NHANES} for background and details.
}
\details{
The National Health and Nutrition Examination Survey (NHANES) is a research survey
distributed to adults and children across the United States to examine health
and nutritional status nationwide. It is unique in that it conducts both interviews
and physical examination to produce their data. The survey is run by the National
Center for Health Statistics (NCHS), a part of the Centers for Disease Control and Prevention.
NA variables means that variable was not recorded for the given observation.
}
\keyword{datasets}
