% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sps.R
\docType{data}
\name{sps}
\alias{sps}
\title{A Randomised Assessment of the Mexican Universal Health Insurance Programme.
(2009)}
\format{
A tibble with 16733 observations and 14 variables: \describe{
\item{age}{numerical variable with raw age values of respondents}
\item{sex}{character variable with values "m" and "f"}
\item{educ}{character variable with values "Preschool", "Primary",
"Secondary", "HS or Vocational", "Normal", "Technical or Commercial", and
"College"} \item{treatment}{binary numerical variable for treatment: 1 =
Treated and 0 = Control} \item{ideology}{character variable with values
"Right", "Center-right", "Center", "Center-left", and "Left"}
\item{ec_better}{factor variable with three levels: "Better", "Same", and "Worse"}
\item{pol_better}{factor variable with three levels: "Better", "Same", and "Worse"}
\item{health_exp_m3}{numerical variable for health-related expenses in the
last 3 month in pesos} \item{health_exp_m1}{numerical variable for
health-related expenses in the last month in pesos}}
}
\source{
\url{https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/P6NC0M}
}
\usage{
sps
}
\description{
Data for a randomised assessment of the Mexican universal health insurance
programme, both before and after a natural experiment which randomly assigned
treatment within 74 matched pairs of health clusters —representing 118 569
households in seven Mexican states. See
\href{https://gking.harvard.edu/files/gking/files/pubpolforpoor.pdf}{King et
al. (2009)} for background and details. measured outcomes in a 2005 baseline
survey (August, 2005, to September, 2005) and follow-up survey 10 months
later (July, 2006, to August, 2006) in 50 pairs (n=32 515)
}
\keyword{datasets}
