% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sps.R
\docType{data}
\name{sps}
\alias{sps}
\title{A Randomised Assessment of the Mexican Universal Health Insurance Programme.
(2009)}
\format{
A tibble with 16,733 observations and 8 variables: \describe{
\item{age}{numerical variable with raw age values of respondents}
\item{sex}{character variable with values "Male" and "Female"}
\item{educ}{character variable with values "Preschool", "Primary",
"Secondary", "High School", "Normal", "Technical",
"College", and "Post-Grad"} \item{treatment}{binary numerical variable for treatment: 1 =
Treated and 0 = Control}
\item{health_exp_m3}{numerical variable for health-related expenses in the
last 3 month in pesos}
\item{health_exp_m1}{numerical variable for
health-related expenses in the last month in pesos}}
}
\source{
\url{https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/P6NC0M}
}
\usage{
sps
}
\description{
Data for a randomised assessment of the Mexican universal health insurance
programme, both before and after a study which randomly assigned
treatment within 74 matched pairs of health clusters --- representing 118,569
households in seven Mexican states. See
\href{https://gking.harvard.edu/files/gking/files/pubpolforpoor.pdf}{King et
al. (2009)} for background and details. Data includes variables from a 2005 baseline
survey (August, 2005, to September, 2005) and follow-up survey 10 months
later (July, 2006, to August, 2006).
}
\keyword{datasets}
